UI changes, RecordingScheduler made thread-safe.

+++ StudioTVPlayer/Providers/RecordingStore.cs
@@ -13,6 +13,9 @@ namespace StudioTVPlayer.Providers
         private readonly string _folder = Path.Combine(GlobalApplicationData.ApplicationDataDir, "Recordings");
         private readonly XmlSerializer _xmlSerializer = XmlSerializer.FromTypes([typeof(Model.Persistence.Recording)])[0];
         private readonly List<Model.Recording> _runningRecordings = [];
+        private List<Model.Recording> _allRecordings;
+        private readonly object _runningRecordingsLock = new();
+        private readonly object _allRecordingsLock = new();
 
         public static RecordingStore Current { get; } = new RecordingStore();
 
@@ -24,62 +27,96 @@ namespace StudioTVPlayer.Providers
             }
         }
 
-        public event EventHandler<RecordingEventArgs> RunningRecordingAdded;
+        public event EventHandler<RecordingEventArgs> RecordingAdded;
+        public event EventHandler<RecordingEventArgs> RecordingDeleted;
 
-        public IReadOnlyList<Model.Recording> RunningRecordings => _runningRecordings;
+        public IEnumerable<Model.Recording> RunningRecordings
+        {
+            get
+            {
+                lock (_runningRecordingsLock)
+                    return [.. _runningRecordings];
+            }
+        }
 
         public void SaveRecording(Model.Recording recording)
         {
             var persistenceRecording = new Model.Persistence.Recording(recording);
-            using (var stream = new FileStream(GetFilePath(recording), FileMode.Create, FileAccess.Write))
-            {
-                _xmlSerializer.Serialize(stream, persistenceRecording);
-            }
+            using var stream = new FileStream(GetRecordingInfoFilePath(recording), FileMode.Create, FileAccess.Write);
+            _xmlSerializer.Serialize(stream, persistenceRecording);
         }
 
         public IEnumerable<Model.Recording> LoadRecordings()
         {
-            return Directory.EnumerateFiles(_folder, "*.xml").Select(fileName =>
-            {
-                using var reader = new StreamReader(fileName);
-                var deserialized = _xmlSerializer.Deserialize(reader) as Model.Persistence.Recording;
-                var runningRecording = _runningRecordings.FirstOrDefault(r => r.Id.ToString() == deserialized.Id);
-                return runningRecording ?? new Model.Recording(deserialized);
-            });
+            lock (_allRecordingsLock)
+                _allRecordings ??= [.. Directory.EnumerateFiles(_folder, "*.xml").Select(fileName =>
+                    {
+                        using var reader = new StreamReader(fileName);
+                        var deserialized = _xmlSerializer.Deserialize(reader) as Model.Persistence.Recording;
+                        Model.Recording runningRecording;
+                        lock (_runningRecordingsLock)
+                        {
+                            runningRecording = _runningRecordings.FirstOrDefault(r => r.Id.ToString() == deserialized.Id);
+                        }
+                        return runningRecording ?? new Model.Recording(deserialized);
+                    })];
+            return [.. _allRecordings];
         }
 
-        public void AddRunningRecording(Model.Recording recording)
+        public void AddRecording(Model.Recording recording)
         {
             Debug.Assert(recording.State is Model.RecordingState.Running);
-            _runningRecordings.Add(recording);
-            RunningRecordingAdded?.Invoke(this, new RecordingEventArgs(recording));
+            lock (_runningRecordingsLock)
+            {
+                _runningRecordings.Add(recording);
+            }
             recording.PropertyChanged += Recording_PropertyChanged;
+            lock (_allRecordingsLock)
+            {
+                DeleteRecording(_allRecordings.FirstOrDefault(r => r.FullPath == recording.FullPath));
+                _allRecordings.Add(recording);
+            }
+            RecordingAdded?.Invoke(this, new RecordingEventArgs(recording));
         }
 
-        internal void DeleteRecording(Model.Recording recording)
+        public void DeleteRecording(Model.Recording recording)
         {
-            var fileName = GetFilePath(recording);
+            if (recording is null)
+                return;
+            lock (_allRecordingsLock)
+            {
+                if (_allRecordings.Remove(recording))
+                    RecordingDeleted?.Invoke(this, new RecordingEventArgs(recording));
+            }
+            var fileName = GetRecordingInfoFilePath(recording);
             if (File.Exists(fileName))
                 File.Delete(fileName);
         }
 
-        private string GetFilePath(Model.Recording recording) => Path.Combine(_folder, $"{recording.Id}.xml");
-        
+        private string GetRecordingInfoFilePath(Model.Recording recording) => Path.Combine(_folder, $"{recording.Id}.xml");
+
         private void Recording_PropertyChanged(object sender, PropertyChangedEventArgs e)
         {
             var recording = sender as Model.Recording ?? throw new ArgumentException(nameof(sender));
             if (e.PropertyName is nameof(Model.Recording.State) && recording.State is Model.RecordingState.Aborted or Model.RecordingState.Completed or Model.RecordingState.Failed)
             {
                 recording.PropertyChanged -= Recording_PropertyChanged;
-                _runningRecordings.Remove(recording);
+                lock (_runningRecordingsLock)
+                {
+                    _runningRecordings.Remove(recording);
+                }
             }
         }
 
         public void Dispose()
         {
-            foreach (var recording in _runningRecordings.ToList())
-                recording.Dispose(); //will also call Recording_Finished below
-
+            IEnumerable<Model.Recording> recordingsToStop;
+            lock (_runningRecordingsLock)
+            {
+                recordingsToStop = [.. _runningRecordings];
+            }
+            foreach (var recording in recordingsToStop)
+                recording.Stop(Model.RecordingState.Aborted); //will also call Recording_Finished below
         }
     }
